Linux 2.2.17pre11
o	Tlan driver update				(Torben Mathiasen)
o	Ext2 error message typo				(Andreas Dilger)
o	Updated Yamaha PCI audio driver			(Daisuke Nagano)
o	Small PPP oops fixes				(Paul Fulghum)
o	SMBFS compile fix				(Urban Widmark)
o	Small cisco hdlc/lmi fixes			(Krzysztof Halasa)
o	Fix hdlc mode switch				(Ivan Passos)
o	Fix fbmem driver bug with size checks		(Petr Vandrovec)
o	Merge the Emu10K driver				(Rui Sousa and co)
o	Synclink driver updates				(Paul Fulghum)
o	Appletalk SMP locking fixes			(Jens-Uwe Mager)
o	Don't offer Toshiba Laptop for non x86		(Francis Galiegue)
o	Further PPC minor merges			(Paul Mackerras)

+++ include/linux/hdlc.h
@@ -157,8 +157,8 @@ typedef struct {
 	u8 N391cnt;
 
 	u8 state;		/* ! */
-	u8 txseq;		/* ! TX sequence number */
-	u8 rxseq;		/* ! RX sequence number */
+	u32 txseq;		/* ! TX sequence number - Cisco uses 4 bytes */
+	u32 rxseq;		/* ! RX sequence number */
 }fr_lmi;			/* ! means used in Cisco HDLC as well */
